fix: WIP refactor HomeScreenContext
