+++ dev-client/src/screens/HomeScreen/HomeScreen.tsx
@@ -16,9 +16,7 @@
  */
 
 import {
-  createContext,
   memo,
-  RefObject,
   useCallback,
   useContext,
   useEffect,
@@ -39,6 +37,7 @@ import {Coords} from 'terraso-client-shared/types';
 import {ListFilterProvider} from 'terraso-mobile-client/components/ListFilter';
 import {Box} from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {useGeospatialContext} from 'terraso-mobile-client/context/GeospatialContext';
+import {HomeScreenContextProvider} from 'terraso-mobile-client/context/HomeScreenContext';
 import {AppBar} from 'terraso-mobile-client/navigation/components/AppBar';
 import {AppBarIconButton} from 'terraso-mobile-client/navigation/components/AppBarIconButton';
 import {LandPKSInfoModal} from 'terraso-mobile-client/screens/HomeScreen/components/LandPKSInfoModal';
@@ -58,24 +57,6 @@ import {getHomeScreenFilters} from 'terraso-mobile-client/screens/HomeScreen/uti
 import {ScreenScaffold} from 'terraso-mobile-client/screens/ScreenScaffold';
 import {useDispatch, useSelector} from 'terraso-mobile-client/store';
 
-type HomeScreenRef = {
-  showSiteOnMap: (site: Site) => void;
-  collapseBottomSheet: () => void;
-};
-
-const HomeScreenContext = createContext<RefObject<HomeScreenRef> | null>(null);
-
-export const HomeScreenContextProvider = memo(
-  ({children}: React.PropsWithChildren<{}>) => (
-    <HomeScreenContext.Provider value={useRef<HomeScreenRef>(null)}>
-      {children}
-    </HomeScreenContext.Provider>
-  ),
-);
-
-export const useHomeScreenContext = () =>
-  useContext(HomeScreenContext)?.current ?? undefined;
-
 export const HomeScreen = memo(() => {
   const infoBottomSheetRef = useRef<BottomSheetModal>(null);
   const siteListBottomSheetRef = useRef<BottomSheet>(null);
@@ -89,7 +70,7 @@ export const HomeScreen = memo(() => {
   const dispatch = useDispatch();
   const mapRef = useRef<MapRef>(null);
   const siteProjectRoles = useSelector(state => selectSitesAndUserRoles(state));
-  const homeScreenContext = useContext(HomeScreenContext);
+  const homeScreenContext = useContext(HomeScreenContextProvider);
 
   const showSiteOnMap = useCallback(
     (targetSite: Site) => {
